+++ doc/CHANGES.md
@@ -9,6 +9,7 @@ Please see [rsbackup in git](https://github.com/ewxrjk/rsbackup) for detailed ch
 * The pruning log in the backup report is now displayed in a less verbose form. Fixes [issue #69](https://github.com/ewxrjk/rsbackup/issues/69).
 * Warnings about unknown hosts and volumes are more informative. Fixes [issue #79](https://github.com/ewxrjk/rsbackup/issues/79).
 * The `rsync --link-dest` option is used in a more subtle way. The most recent complete backup is always used if possible, but if there are more recent incomplete backups then the latest of those is used as well. Fixes [issue #77](https://github.com/ewxrjk/rsbackup/issues/77).
+* The `decay-scale` prune parameter is no longer constrained to be an integer. Fixes [issue #87](https://github.com/ewxrjk/rsbackup/issues/87).
 
 ## Changes In rsbackup 7.0